requirements.txt
numpy
torch
pyyaml